Avoid using undefined buffer contents in bf_ctime()
The new bf_ctime() with optional timezone argument wasn't checking
the return value from strftime(), potentially leaving the result
buffer with undefined contents if e.g. a pathologically long timezone
string is passed in.

Perhaps there could be better verification that the timezone is
indeed valid, but for now we truncate the buffer and raise E_INVARG
when strftime indicates there's not enough space for the result.