Improve stream robustness
The stream code assumes it can append a null byte before returning its
contents to the client, but it allows the stream to be initialized with
a size of zero.  Fortunately mymalloc() saves us by allocating one byte
rather than none at all, but we should not depend on this.  Also, a
size of zero means any attempt to grow by s->buflen * 2 will fail.