Enhance encode_chars() and decode_chars()
Now the first argument to encode_chars() can be a list of strings and/or
integer character codes, and decode_chars() returns a list of same, for
better symmetry with encode_binary()/decode_binary().

The second argument <encoding> to both functions is now mandatory.

A third optional argument <fully> to decode_chars() forces all characters
to be returned as integer character codes rather than as a string.