Clean up and add configure checks
The option --without-included-pcre can be given to `configure' to use the
system's installed PCRE library rather than the included one, as long as it
was built with support for both UTF-8 and Unicode character properties.

In pattern.c, debug checks have been turned off, resource limits have been
changed to approximate the previous implementation's, and the code is
otherwise minorly cleaned up.