Update bf_random() to support 64-bit numbers
The MOO random() built-in function failed to play nice with 64-bit values,
for example raising E_INVARG whenever an argument greater than 2^31-1 was
passed.  Also, it was apparently possible for random() with no argument to
return the value 0, when it is documented as returning a number in the
range [1..$maxint].  Furthermore, the implementation made assumptions about
the value of RAND_MAX which may be inappropriate.

This version allows random() to return any value in the range [1..2^63-1]
independent of the value of RAND_MAX.