Fix random(m) to be uniformly distributed for m!=2^k
Also allows for m > RAND_MAX and includes generalizations
for 64-bit environments as on the unicode(64) branch.
(for now this code is confined to numbers.c; see comment there)

In order for RANDOM()%m to be distributed uniformly,
the range of RANDOM() has to be divisible by m.
A typical screw case is:

    n = 0;
    m = $maxint/3;
    for i in [1..10000]
      if (random(2*m) > m)
	n = n+1;
      endif
    endfor
    return n;

which should return something between 4900-5100,
prior version would return ~3333.